yaml2wiki/requirements.txt
pyyaml
wikitools
